power_path: split into early and late setup
Configuration of power path is crutial to start battery charging. When
the battery is deeply discharged to a degree that the fuel gauge has gone
into under voltage lockout the charger needs to be engaged before
communication with the gauge wull be possible again.
Initialize battery charger before trying to communicate with battery
gauge.