i2c_bus: support unsticking locked up I2C bus
Simple bare I2C devices (like for example an LM75) can lock up I2C bus
communications.
This can happen if due to software error or EMI event a clock pulse is
missed while reading from an I2C secondary device. If the I2C device
happens to have been transmitting a zero bit at the time of the
disturbance and this is the last bit sent it may never release the bus
since - from the perspective of the device - there is still one more bit
to be sent.
Consequently the primary device will not be able to start new bus
transfers since a start condition can no longer be signaled.
Send dummy clock pulses on the SCL line if any such condition is detected
to unstick the bus.